+++ src/euphorie/client/survey.py
@@ -179,8 +179,14 @@ class Resume(grok.View):
             # No tree generated, so start over
             self.request.response.redirect("%s/start" % survey.absolute_url())
         else:
+            # Redirect to the start page of the Identification phase.
+            # We do this to ensure the screen with the tool name gets shown.
+            # If we jump directly to the first question, the user does not
+            # see the tool name.
+            # This is especially relevant since the osc-header now displays the
+            # user-given session name.
             self.request.response.redirect(
-                    QuestionURL(survey, question, phase="identification"))
+                "{0}/{1}".format(survey.absolute_url(), "identification"))
 
 
 class Identification(grok.View):
@@ -215,6 +221,23 @@ class Identification(grok.View):
             self.tree = getTreeData(self.request, question)
         else:
             self.next_url = None
+
+    @property
+    def extra_text(self):
+        appconfig = getUtility(IAppConfig)
+        settings = appconfig.get('euphorie')
+        have_extra = settings.get('extra_text_idendification', False)
+        if not have_extra:
+            return None
+        lang = getattr(self.request, 'LANGUAGE', 'en')
+        # Special handling for Flemish, for which LANGUAGE is "nl-be". For
+        # translating the date under plone locales, we reduce to generic "nl".
+        # For the specific oira translation, we rewrite to "nl_BE"
+        if "-" in lang:
+            elems = lang.split("-")
+            lang = "{0}_{1}".format(elems[0], elems[1].upper())
+        return translate(_(
+            u"extra_text_identification", default=u""), target_language=lang)
 
 
 class ActionPlan(grok.View):
@@ -290,6 +313,11 @@ class Status(grok.View):
         )
         self.label_page = translate(_(u"label_page", default=u"Page"), target_language=lang)
         self.label_page_of = translate(_(u"label_page_of", default=u"of"), target_language=lang)
+        session = SessionManager.session
+        if session is not None and session.title != self.context.Title():
+            self.session_title = session.title
+        else:
+            self.session_title = None
 
     def module_query(self, sessionid, optional_modules):
         if optional_modules: